CR feedback - $gridaction -> $grid-action - move Change and ChangeType to /grid/model.ts
