+++ packages/components/src/public/QueryModel/grid/model.ts
@@ -1,0 +1,12 @@
+
+export enum ChangeType {
+    add = 'add',
+    remove = 'remove',
+    modify = 'modify',
+    none = 'none',
+}
+
+export interface Change {
+    type: ChangeType;
+    index?: number;
+}
