+++ packages/components/src/theme/query-model.scss
@@ -96,10 +96,10 @@
         line-height: 16px;
         margin-right: 8px;
         padding: 8px 8px 8px 32px;
-        color: $gridaction-item-color;
-        background-color: $gridaction-item-bg;
+        color: $grid-action-item-color;
+        background-color: $grid-action-item-bg;
         border-radius: 2px;
-        border: 1px solid $gridaction-item-border-color;
+        border: 1px solid $grid-action-item-border-color;
 
         i.read-lock {
             padding-right: 6px;
@@ -107,38 +107,38 @@
         }
 
         i.symbol {
-            color: $gridaction-item-color;
+            color: $grid-action-item-color;
             width: 23px;
             height: 14px;
             margin-right: 8px;
             margin-left: -25px;
-            border-right: solid 1px $gridaction-item-border-color;
+            border-right: solid 1px $grid-action-item-border-color;
         }
 
         &:hover,
         &:focus {
-            background-color: $gridaction-item-hover-bg;
+            background-color: $grid-action-item-hover-bg;
         }
         &:active {
-            background-color: $gridaction-item-border-color;
+            background-color: $grid-action-item-border-color;
         }
 
         &.is-disabled {
-            background-color: $gridaction-item-disabled-bg;
-            border: 1px solid $gridaction-item-disabled-border-color;
-            color: $gridaction-item-disabled-color;
+            background-color: $grid-action-item-disabled-bg;
+            border: 1px solid $grid-action-item-disabled-border-color;
+            color: $grid-action-item-disabled-color;
 
             i.symbol {
-                color: $gridaction-item-disabled-color;
-                border-color: $gridaction-item-disabled-border-color;
+                color: $grid-action-item-disabled-color;
+                border-color: $grid-action-item-disabled-border-color;
             }
 
             &:hover,
             &:focus {
-                background-color: $gridaction-item-disabled-bg;
+                background-color: $grid-action-item-disabled-bg;
             }
             &:active {
-                background-color: $gridaction-item-disabled-bg;
+                background-color: $grid-action-item-disabled-bg;
             }
         }
     }
